Set first click event to null
Only for deco spot controllers

this is to make sure that first click event does not get called after one of them has been clicked. Otherwise, the start menu panel will be turn off then on again.

The rest of the base class event will still be called. So the audio clip, which will be set in the future, should still be played however many times the deco spots have been pressed.