UI animation has its own base class
This is due to the nature that most of the targets that are used in cases where the control is a UI Element, are mostly set in the Inspector. Unlike Move To or RotateTowards, the target values for UI's are not serialized at runtime in script, and are most suitable for examination in Inspector anyway.
Future reform should focus on creating a deeper base class that both AAnimation and AUIAnimation can inherit from while both use the same PerfromAsync without breaking polymorphism.