Fixed SettingsManager.cs nullref
Checks if CurrentToggleController exits before referencing it in the SaveSettings method.

Only the Theme tab needs this kind of check. This is because the General tab will always the shown as the first tab. Therefore, the ToggleController's used in that tab will always be initialized and thus CurrentToggleController of the ToggleControllerGroup that sits inside the General Tab will never be null.