Fixed bug
Previously when selecting a different theme in the Settings panel, the settings data were saved. However when SpotMaterialController tried to access the CurrentTheme in  SettingsManager, it will always get the older choice of theme. This only got fixed when entering the settings panel again.

This was due to the fact that CurrentTheme relied on _themeID as the index to get the theme. The problem was that _themeID was never updated when we saved to the json file.

Now the CurrentTheme uses MasterSettingsData.ThemeSettingsData.ThemeID, which would be updated when exiting settings panel.

Also, _themeID was renamed to _oldThemeID. This is a better name since we only use it when trying to load FROM settings or when some components are null and to use them as safeguard.